:truck: create the new trackPageActivities module and it generic waitIdlePageActivity function
